==> AM-DNA-258/5_Itol_Grapher.ipynb <==
from IPython.display import Image
#import gurobipy
import cassiopeia as cas
import pandas as pd
import numpy as np
import pickle
import os
import matplotlib.pyplot as plt
import seaborn as sns
from ete3 import Tree
from typing import Tuple
import scipy

start_dir = '/Genomics/chanlab/blaw/TLS/data/AM-DNA-258/lineage/3_lineage_reconstruction/'
out_dir = '/Genomics/chanlab/blaw/TLS/data/AM-DNA-258/lineage/tree_graphing/'

clusterColorsFile = "/Genomics/chanlab/mchan/Adriano/TLS/TLS_TLSCL/20211102_clusterColorsTLSCL.p"
with open(clusterColorsFile,'rb') as fp:
    colorDict = pickle.load(fp)

colors = ['#e6194b', '#3cb44b', '#ffe119', '#4363d8', '#f58231', '#911eb4', '#46f0f0', '#f032e6', '#bcf60c', '#fabebe', 
          '#008080', '#e6beff', '#9a6324', '#fffac8', '#800000', '#aaffc3', '#808000', '#ffd8b1', '#000075', '#808080', 
          '#ffffff', '#000000', '#1f77b4', '#aec7e8', '#ff7f0e', '#ffbb78', '#2ca02c', '#98df8a', '#d62728', '#ff9896',
          '#9467bd', '#c5b0d5', '#8c564b', '#c49c94', '#e377c2', '#f7b6d2', '#7f7f7f', '#c7c7c7', '#bcbd22', '#dbdb8d',
          '#17becf', '#9edae5']

def plotItol (barcode, file, colorList):
    t = Tree(start_dir + barcode + '/' + file, format=1)
    tree_meta = pd.read_csv(start_dir + barcode + '/' + barcode + '_metadata.txt', sep='\t')
    character_matrix = pd.read_csv(start_dir + barcode + '/' + barcode + '_character_matrix.txt', sep='\t', index_col = 0)

    prior_file = start_dir + barcode + '/' + barcode + '_priors.pickle'
    with open(prior_file, 'rb') as f:
        priors = pickle.load(f)

    test_tree = cas.data.CassiopeiaTree(character_matrix=character_matrix, priors=priors, tree = t)

    missing_proportion = (character_matrix == -1).sum(axis=0) / character_matrix.shape[0]
    uncut_proportion = (character_matrix == 0).sum(axis=0) / character_matrix.shape[0]
    n_unique_states = character_matrix.apply(lambda x: len(np.unique(x[(x != 0) & (x != -1)])), axis=0)

    character_meta = pd.DataFrame([missing_proportion, uncut_proportion, n_unique_states], index = ['missing_prop', 'uncut_prop', 'n_unique_states']).T
    test_tree.cell_meta = tree_meta
    test_tree.character_meta = character_meta

    test_tree.cell_meta['cluster'] = test_tree.cell_meta['cell_state'].str[2:-2]
    test_tree.cell_meta['lenti'] = test_tree.cell_meta['StF_Group'].str[2:-2]
    test_tree.cell_meta.set_index("cellBC",inplace=True)

    cas.pl.upload_and_export_itol(test_tree, 'AM-DNA-258_' + barcode + '_Final', api_key = 'jtMaYpxAc94ImqU7xJjubQ',
                                  project_name = 'AM-DNA-258', export_filepath= out_dir + barcode + '_Final_Lenti_Tree.pdf', 
                                  meta_data = ['cluster', 'lenti'], palette = colorList)
    
    return test_tree

Bar1_color = ['#0000cd', '#30D5C8',
             '#228B22', '#023020',
             '#ff0000', '#ba55d3',
             '#800080', '#9400d3',
             '#FF00FF', '#FFC0CB',
             '#808080', '#483d8b',
             '#6a5acd', '#FFFF00']

file = 'Bar1_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar1'

test_tree = plotItol(barcode, file, Bar1_color)

Bar2_color = ['#0000cd', '#30D5C8',
             '#228B22', '#023020',
             '#ba55d3', '#800080',
             '#9400d3', '#FF00FF',
             '#FFC0CB', '#808080',
             '#483d8b', '#6a5acd',
             '#FFFF00', '#FF00FF']

file = 'Bar2_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar2'

test_tree = plotItol(barcode, file, Bar2_color)

Bar3_color = ['#30D5C8', '#228B22',
              '#ba55d3', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '#483d8b',
              '#6a5acd', '',
              '', '',
              '', '']

file = 'Bar3_hybrid_newick_noMutationlessEdges.txt'
barcode = 'Bar3'

test_tree = plotItol(barcode, file, colors)

Bar4_color = ['#0000cd', '#30D5C8',
             '#228B22', '#023020',
             '#ba55d3', '#800080',
             '#9400d3', '#FF00FF',
             '#FFC0CB', '#808080',
             '#483d8b', '#6a5acd',
             '#FFFF00', '#ff0000']

file = 'Bar4_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar4'

test_tree = plotItol(barcode, file, Bar4_color)

Bar5_color = ['#FFFF00', '#0000cd',
             '#30D5C8', '#228B22',
             '#023020', '#ff0000',
             '#ba55d3', '#800080',
             '#9400d3', '#FF00FF',
             '#FFC0CB', '#30D5C8',
             '#483d8b', '#6a5acd']

file = 'Bar5_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar5'

test_tree = plotItol(barcode, file, Bar5_color)

Bar6_color = ['#FFFF00', '#30D5C8',
             '#30D5C8', '#ba55d3',
             '#9400d3', '#FF00FF',
             '#FFC0CB', '#483d8b',
             '#6a5acd', '#FF00FF',
             '#FFC0CB', '#228B22',
             '#483d8b', '#6a5acd']

file = 'Bar6_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar6'

test_tree = plotItol(barcode, file, Bar6_color)

Bar7_color = ['#30D5C8', '#228B22',
                '#023020', '#ff0000',
                '#ba55d3', '#800080',
                '#9400d3', '#FF00FF',
                '#FFC0CB', '#808080',
                '#483d8b', '#6a5acd',
                '#0000cd', '']

file = 'Bar7_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar7'

test_tree = plotItol(barcode, file, Bar7_color)

Bar9_color = ['#30D5C8', '#228B22',
              '#023020', '#ff0000',
              '#ba55d3', '#800080',
              '#9400d3', '#FF00FF',
              '#FFC0CB', '#483d8b',
              '#6a5acd', '#FFFF00',
              '#0000cd', '#808080']

file = 'Bar9_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar9'

test_tree = plotItol(barcode, file, Bar9_color)

Bar10_color = ['#FFFF00', '#FFC0CB',
              '#30D5C8', '#228B22',
              '#023020', '#ba55d3',
              '#800080', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '#483d8b',
              '#6a5acd', '#ff0000']

file = 'Bar10_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar10'

test_tree = plotItol(barcode, file, Bar10_color)

Bar11_color = ['#0000cd', '#30D5C8',
              '#228B22', '#023020',
              '#ff0000', '#ba55d3',
              '#800080', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#483d8b', '#6a5acd',
              '#808080', '#FFFF00']

file = 'Bar11_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar11'

test_tree = plotItol(barcode, file, Bar11_color)

Bar12_color = ['#FFFF00', '#30D5C8',
              '#228B22', '#023020',
              '#ff0000', '#ba55d3',
              '#800080', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#483d8b', '#6a5acd',
              '#808080', '#0000cd']

file = 'Bar12_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar12'

test_tree = plotItol(barcode, file, Bar12_color)

Bar13_color = ['#0000cd', '#30D5C8',
              '#228B22', '#023020',
              '#ba55d3', '#800080',
              '#9400d3', '#FF00FF',
              '#FFC0CB', '#808080',
              '#483d8b', '#6a5acd',
              '#ff0000', '#FFFF00']

file = 'Bar13_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar13'

test_tree = plotItol(barcode, file, Bar13_color)

Bar14_color = ['#FFFF00', '#30D5C8',
              '#228B22', '#ba55d3',
              '#800080', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '#483d8b',
              '#6a5acd', '#023020',
              '#0000cd', '#ff0000']

file = 'Bar14_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar14'

test_tree = plotItol(barcode, file, colors)

Bar15_color = ['#30D5C8', '#228B22',
              '#ba55d3', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '#483d8b',
              '#6a5acd', '',
              '', '',
              '', '']

file = 'Bar15_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar15'

test_tree = plotItol(barcode, file, Bar15_color)

Bar16_color = ['#FFFF00', '#0000cd',
              '#30D5C8', '#228B22',
              '#023020', '#ba55d3',
              '#800080', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '#483d8b',
              '#6a5acd', '']

file = 'Bar16_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar16'

test_tree = plotItol(barcode, file, Bar16_color)

Bar18_color = ['#ba55d3', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '',
              '', '',
              '', '',
              '', '',
              '', '']

file = 'Bar18_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar18'

test_tree = plotItol(barcode, file, Bar18_color)

Bar19_color = ['#0000cd', '#30D5C8',
              '#228B22', '#023020',
              '#ba55d3', '#800080',
              '#9400d3', '#FF00FF',
              '#FFC0CB', '#808080',
              '#483d8b', '#6a5acd',
              '', '']

file = 'Bar19_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar19'

test_tree = plotItol(barcode, file, Bar19_color)

Bar20_color = ['#30D5C8', '#228B22',
              '#ba55d3', '#800080',
              '#9400d3', '#FF00FF',
              '#FFC0CB', '#808080',
              '#483d8b', '#6a5acd',
              '', '',
              '', '']

file = 'Bar20_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar20'

test_tree = plotItol(barcode, file, Bar20_color)

Bar21_color = ['#30D5C8', '#228B22',
              '#ff0000', '#ba55d3',
              '#800080', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '#483d8b',
              '#6a5acd', '',
              '', '']

file = 'Bar21_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar21'

test_tree = plotItol(barcode, file, Bar21_color)

Bar22_color = ['#FFFF00', '#0000cd',
              '#30D5C8', '#228B22',
              '#023020', '#ff0000',
              '#ba55d3', '#800080',
              '#9400d3', '#FF00FF',
              '#FFC0CB', '#808080',
              '#483d8b', '#6a5acd']

file = 'Bar22_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar22'

test_tree = plotItol(barcode, file, Bar22_color)

Bar23_color = ['#30D5C8', '#228B22',
              '#ba55d3', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#808080', '#483d8b',
              '#6a5acd', '',
              '', '',
              '', '']

file = 'Bar23_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar23'

test_tree = plotItol(barcode, file, Bar23_color)

Bar24_color = ['#FFFF00', '#30D5C8',
              '#023020', '#ba55d3',
              '#800080', '#9400d3',
              '#FF00FF', '#FFC0CB',
              '#483d8b', '#6a5acd',
              '', '',
              '', '']

file = 'Bar24_ilp_newick_noMutationlessEdges.txt'
barcode = 'Bar24'

test_tree = plotItol(barcode, file, Bar24_color)
